Add mesos master allocation metrics.
I reviewed the metrics documented here:

http://mesos.apache.org/documentation/latest/monitoring/#allocator

I elided metrics which had a variable <task>, <role>, etc because
I wasn't sure how to approach that.  If we want to try to
parse them out from the map of collected metrics and add those
as labels, we can do that, but I'd probably need some direction
there.